Rename `resolve` package to `converters`
