+++ compiler/ballerina-lang/src/main/java/org/wso2/ballerinalang/compiler/packaging/converters/Converter.java
@@ -1,4 +1,4 @@
-package org.wso2.ballerinalang.compiler.packaging.resolve;
+package org.wso2.ballerinalang.compiler.packaging.converters;
 
 
 import java.nio.file.Path;
