+++ compiler/ballerina-lang/src/main/java/org/wso2/ballerinalang/compiler/packaging/repo/CacheRepo.java
@@ -2,8 +2,8 @@ package org.wso2.ballerinalang.compiler.packaging.repo;
 
 import org.ballerinalang.model.elements.PackageID;
 import org.wso2.ballerinalang.compiler.packaging.Patten;
-import org.wso2.ballerinalang.compiler.packaging.resolve.Converter;
-import org.wso2.ballerinalang.compiler.packaging.resolve.PathConverter;
+import org.wso2.ballerinalang.compiler.packaging.converters.Converter;
+import org.wso2.ballerinalang.compiler.packaging.converters.PathConverter;
 
 import java.nio.file.Path;
