+++ compiler/ballerina-lang/src/main/java/org/wso2/ballerinalang/compiler/packaging/repo/Repo.java
@@ -2,7 +2,7 @@ package org.wso2.ballerinalang.compiler.packaging.repo;
 
 import org.ballerinalang.model.elements.PackageID;
 import org.wso2.ballerinalang.compiler.packaging.Patten;
-import org.wso2.ballerinalang.compiler.packaging.resolve.Converter;
+import org.wso2.ballerinalang.compiler.packaging.converters.Converter;
 
 /**
  * Calculate the path patten of a given package.
